requirements.txt
telethon
pandas
python-dotenv
matplotlib
wordcloud
nltk
